refactor: detach duplicate code
