+++ src/component/scatterSeries.ts
@@ -3,9 +3,8 @@ import { CircleModel } from '@t/components/series';
 import { CoordinateDataType, ScatterChartOptions, ScatterSeriesType } from '@t/options';
 import { ClipRectAreaModel } from '@t/components/series';
 import { ChartState, SeriesTheme, ValueEdge } from '@t/store/store';
-import { isNumber } from '@src/helpers/utils';
 import { TooltipData } from '@t/components/tooltip';
-import { hoverStyle } from '@src/brushes/basic';
+import { getCoordinateDataIndex, getCoordinateValue } from '@src/helpers/coordinate';
 
 type DrawModels = ClipRectAreaModel | CircleModel;
 
@@ -29,24 +28,6 @@ export default class ScatterSeries extends Component {
     if (this.models[0].type === 'clipRectArea') {
       this.models[0].width = this.rect.width * delta;
     }
-  }
-
-  getValue(datum: number | CoordinateDataType) {
-    if (isNumber(datum)) {
-      return datum;
-    }
-
-    return Array.isArray(datum) ? datum[1] : datum.y;
-  }
-
-  getDataIndex(datum: number | CoordinateDataType, categories: string[], dataIndex: number) {
-    if (isNumber(datum)) {
-      return dataIndex;
-    }
-
-    const value = Array.isArray(datum) ? datum[0] : datum.x;
-
-    return categories.findIndex(category => category === String(value));
   }
 
   render(chartState: ChartState<ScatterChartOptions>) {
@@ -78,12 +59,12 @@ export default class ScatterSeries extends Component {
     const tooltipDataArr = scatterData.flatMap(({ data, name }, index) => {
       const tooltipData: TooltipData[] = [];
 
-      data.forEach((datum: CoordinateDataType | number, dataIdx) => {
+      data.forEach((datum: CoordinateDataType, dataIdx) => {
         tooltipData.push({
           label: name,
           color: theme.series.colors[index],
-          value: this.getValue(datum),
-          category: categories[this.getDataIndex(datum, categories, dataIdx)]
+          value: getCoordinateValue(datum),
+          category: categories[getCoordinateDataIndex(datum, categories, dataIdx)]
         });
       });
 
@@ -93,7 +74,7 @@ export default class ScatterSeries extends Component {
     this.models = [this.renderClipRectAreaModel(), ...seriesModel];
     this.responders = seriesModel.map((m, index) => ({
       ...m,
-      style: { ...m.style, ...hoverStyle },
+      style: { ...m.style, shadowColor: 'rgba(0, 0, 0, 0.5)', shadowBlur: 6, shadowOffsetY: 2 },
       data: tooltipDataArr[index]
     }));
   }
@@ -120,22 +101,23 @@ export default class ScatterSeries extends Component {
 
     return seriesRawData.flatMap(({ data }, seriesIndex) => {
       const circleModels: CircleModel[] = [];
+      const style = {
+        color: colors[seriesIndex],
+        radius: 7,
+        strokeStyle: '#fff',
+        globalAlpha: 0.8
+      };
 
       data.forEach((datum, idx) => {
-        const value = this.getValue(datum);
-        const dataIndex = this.getDataIndex(datum, categories, idx);
+        const value = getCoordinateValue(datum);
+        const dataIndex = getCoordinateDataIndex(datum, categories, idx);
 
         const valueRatio = (value - limit.min) / (limit.max - limit.min);
 
         const x = tickDistance * dataIndex;
         const y = (1 - valueRatio) * this.rect.height;
 
-        circleModels.push({
-          x,
-          y,
-          type: 'circle',
-          style: { color: colors[seriesIndex], radius: 7, strokeStyle: '#fff', globalAlpha: 0.8 }
-        });
+        circleModels.push({ x, y, type: 'circle', style });
       });
 
       return circleModels;
