+++ types/components/series.d.ts
@@ -1,17 +1,19 @@
 import { Point, Rect, BezierPoint } from '../options';
 
+export interface CircleStyle {
+  color: string;
+  radius: number;
+  globalAlpha?: number;
+  strokeStyle?: string;
+  lineWidth?: number;
+  shadowColor?: string;
+  shadowBlur?: number;
+  shadowOffsetY?: number;
+}
+
 export type CircleModel = {
   type: 'circle';
-  style: {
-    color: string;
-    radius: number;
-    globalAlpha?: number;
-    strokeStyle?: string;
-    lineWidth?: number;
-    shadowColor?: string;
-    shadowBlur?: number;
-    shadowOffsetY?: number;
-  };
+  style: CircleStyle;
 } & Point;
 
 export type ClipRectAreaModel = {
